reorganization and general cleanup. Added some comments
