+++ src/expectation/problem_types.jl
@@ -2,21 +2,24 @@
 abstract type AbstractUncertaintyProblem end
 
 struct ExpectationProblem{TS, TG, TH, TF, TP} <: AbstractUncertaintyProblem
-    # ∫ g(S(h(x,p)))*f(x)dx
-    S::TS  # mapping,                 S: 𝕐 × ℚ → 𝕐
-    g::TG  # observable(output_func), g: 𝕐 × ℚ → ℝⁿᵒᵘᵗ  #TODO or should this be g: 𝕐 × ℙ → ℝⁿᵒᵘᵗ, requires update in build_integrand, if so
-    h::TH  # cov(input_func),         h: 𝕏 × ℙ → 𝕐 × ℚ
+    # defines ∫ g(S(h(x,u0,p)))*f(x)dx
+    # 𝕏 = uncertainty space, 𝕌 = Initial condition space, ℙ = model parameter space,
+    S::TS  # mapping,                 S: 𝕌 × ℙ → 𝕌
+    g::TG  # observable(output_func), g: 𝕌 × ℙ → ℝⁿᵒᵘᵗ
+    h::TH  # cov(input_func),         h: 𝕏 × 𝕌 × ℙ → 𝕌 × ℙ
     d::TF  # distribution,            pdf(d,x): 𝕏 → ℝ
     params::TP
     nout::Int
 end 
 
+# Constructor for general maps/functions
 function ExpectationProblem(g, pdist, params; nout = 1)
     h(x,u,p) = x, p
     S(x,p) = x
     ExpectationProblem(S, g, h, pdist, params, nout)
 end
 
+# Constructor for DEProblems
 function ExpectationProblem(sm::SystemMap, g, h, d; nout = 1)
     ExpectationProblem(sm, g, h, d, 
         ArrayPartition(deepcopy(sm.prob.u0),deepcopy(sm.prob.p)),
@@ -29,6 +32,7 @@ observable(prob::ExpectationProblem) = prob.g
 input_cov(prob::ExpectationProblem) = prob.h
 parameters(prob::ExpectationProblem) = prob.params
 
+# Builds integrand for DEProblems
 function build_integrand(prob::ExpectationProblem{F}) where F<:SystemMap
     @unpack S, g, h, d = prob
     function(x,p)
@@ -37,6 +41,7 @@ function build_integrand(prob::ExpectationProblem{F}) where F<:SystemMap
     end
 end
 
+# Builds integrand for arbitrary functions
 function build_integrand(prob::ExpectationProblem)
     @unpack g, d = prob
     function(x,p)
